+++ acceptance-tests/src/acceptance-test/java/tech/pegasys/teku/test/acceptance/BellatrixMergeTransitionAcceptanceTest.java
@@ -13,6 +13,11 @@
 
 package tech.pegasys.teku.test.acceptance;
 
+import com.google.common.io.Resources;
+import java.io.IOException;
+import java.net.URL;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import tech.pegasys.teku.infrastructure.time.SystemTimeProvider;
@@ -23,6 +28,7 @@ import tech.pegasys.teku.test.acceptance.dsl.TekuNode;
 import tech.pegasys.teku.test.acceptance.dsl.tools.deposits.ValidatorKeystores;
 
 public class BellatrixMergeTransitionAcceptanceTest extends AcceptanceTestBase {
+  private static final Logger LOG = LogManager.getLogger();
   private static final String NETWORK_NAME = "less-swift";
 
   private final SystemTimeProvider timeProvider = new SystemTimeProvider();
@@ -40,17 +46,24 @@ public class BellatrixMergeTransitionAcceptanceTest extends AcceptanceTestBase {
     final int totalValidators = 4;
     final ValidatorKeystores validatorKeystores =
         createTekuDepositSender(NETWORK_NAME).sendValidatorDeposits(eth1Node, totalValidators);
+    final URL jwtFile = Resources.getResource("teku/ee-jwt-secret.hex");
 
     tekuNode =
         createTekuNode(
-            config ->
+            config -> {
+              try {
                 configureTekuNode(config, genesisTime)
                     .withDepositsFrom(eth1Node)
                     .withStartupTargetPeerCount(0)
                     .withValidatorKeystores(validatorKeystores)
                     .withValidatorProposerDefaultFeeRecipient(
                         "0xfe3b557e8fb62b89f4916b721be55ceb828dbd73")
-                    .withExecutionEngineEndpoint(eth1Node.getInternalEngineJsonRpcUrl()));
+                    .withExecutionEngineEndpoint(eth1Node.getInternalEngineJsonRpcUrl())
+                    .withJwtSecretFile(jwtFile.openStream());
+              } catch (IOException e) {
+                LOG.error("JWT secret file loading failure", e);
+              }
+            });
     tekuNode.start();
   }