+++ pow/src/main/java/tech/pegasys/teku/pow/Web3jEth1Provider.java
@@ -58,20 +58,17 @@ public class Web3jEth1Provider extends AbstractMonitorableEth1Provider {
   private final Web3j web3j;
   private final AsyncRunner asyncRunner;
   private final LabelledMetric<Counter> requestCounter;
-  private final boolean trustedWeb3j;
 
   public Web3jEth1Provider(
       final SpecConfig config,
       final MetricsSystem metricsSystem,
       final String id,
       final Web3j web3j,
-      final boolean trustedWeb3j,
       final AsyncRunner asyncRunner,
       final TimeProvider timeProvider) {
     super(timeProvider);
     this.config = config;
     this.web3j = web3j;
-    this.trustedWeb3j = trustedWeb3j;
     this.asyncRunner = asyncRunner;
     this.id = id;
     requestCounter =
@@ -216,22 +213,14 @@ public class Web3jEth1Provider extends AbstractMonitorableEth1Provider {
   public SafeFuture<Boolean> validate() {
     if (validating.compareAndSet(false, true)) {
       LOG.debug("Validating endpoint {} ...", this.id);
-      final SafeFuture<Result> firstValidationStep;
-      if (trustedWeb3j) {
-        firstValidationStep = validateSyncing();
-      } else {
-        firstValidationStep =
-            validateChainId()
-                .thenCompose(
-                    result -> {
-                      if (result == Result.FAILED) {
-                        return SafeFuture.completedFuture(result);
-                      } else {
-                        return validateSyncing();
-                      }
-                    });
-      }
-      return firstValidationStep
+      return validateChainId()
+          .thenCompose(
+              result -> {
+                if (result == Result.FAILED) {
+                  return SafeFuture.completedFuture(result);
+                }
+                return validateSyncing();
+              })
           .thenApply(
               result -> {
                 updateLastValidation(result);
@@ -255,8 +244,13 @@ public class Web3jEth1Provider extends AbstractMonitorableEth1Provider {
 
   private SafeFuture<Result> validateChainId() {
     return getChainId()
-        .thenApply(
-            chainId -> {
+        .handle(
+            (chainId, ex) -> {
+              if (ex != null) {
+                LOG.info(
+                    "eth_chainId method is not supported by provider, skipping validation", ex);
+                return Result.NOT_SUPPORTED;
+              }
               if (chainId.intValueExact() != config.getDepositChainId()) {
                 STATUS_LOG.eth1DepositChainIdMismatch(
                     config.getDepositChainId(), chainId.intValueExact(), this.id);