+++ teku/src/main/java/tech/pegasys/teku/services/BeaconNodeServiceController.java
@@ -15,11 +15,11 @@ package tech.pegasys.teku.services;
 
 import java.util.Optional;
 import tech.pegasys.teku.config.TekuConfiguration;
+import tech.pegasys.teku.infrastructure.el.ExecutionClientProvider;
 import tech.pegasys.teku.networking.nat.NatService;
 import tech.pegasys.teku.service.serviceutils.ServiceConfig;
 import tech.pegasys.teku.services.beaconchain.BeaconChainService;
 import tech.pegasys.teku.services.chainstorage.StorageService;
-import tech.pegasys.teku.services.executionengine.ExecutionClientProvider;
 import tech.pegasys.teku.services.executionengine.ExecutionEngineService;
 import tech.pegasys.teku.services.powchain.PowchainService;
 import tech.pegasys.teku.validator.client.ValidatorClientService;
@@ -30,16 +30,19 @@ public class BeaconNodeServiceController extends ServiceController {
       TekuConfiguration tekuConfig, final ServiceConfig serviceConfig) {
     // Note services will be started in the order they are added here.
     services.add(new StorageService(serviceConfig, tekuConfig.storageConfiguration()));
-    final ExecutionClientProvider trustedExecutionClientProvider;
+    Optional<ExecutionClientProvider> maybeExecutionClientProvider = Optional.empty();
     if (tekuConfig.executionEngine().isEnabled()) {
       // Need to make sure the execution engine is listening before starting the beacon chain
-      trustedExecutionClientProvider =
-          ExecutionClientProvider.create(serviceConfig, tekuConfig.executionEngine());
+      final ExecutionClientProvider executionClientProvider =
+          ExecutionClientProvider.create(
+              tekuConfig.executionEngine().getEndpoint(),
+              serviceConfig.getTimeProvider(),
+              tekuConfig.executionEngine().getJwtSecretFile(),
+              serviceConfig.getDataDirLayout().getBeaconDataDirectory());
       services.add(
           new ExecutionEngineService(
-              serviceConfig, tekuConfig.executionEngine(), trustedExecutionClientProvider));
-    } else {
-      trustedExecutionClientProvider = ExecutionClientProvider.NOOP;
+              serviceConfig, tekuConfig.executionEngine(), executionClientProvider));
+      maybeExecutionClientProvider = Optional.of(executionClientProvider);
     }
     services.add(new BeaconChainService(serviceConfig, tekuConfig.beaconChain()));
     services.add(
@@ -47,25 +50,21 @@ public class BeaconNodeServiceController extends ServiceController {
             tekuConfig.natConfiguration(),
             tekuConfig.network().getListenPort(),
             tekuConfig.discovery().isDiscoveryEnabled()));
-    powchainService(tekuConfig, serviceConfig, trustedExecutionClientProvider)
+    powchainService(tekuConfig, serviceConfig, maybeExecutionClientProvider)
         .ifPresent(services::add);
     services.add(ValidatorClientService.create(serviceConfig, tekuConfig.validatorClient()));
   }
 
   private Optional<PowchainService> powchainService(
-      TekuConfiguration tekuConfig,
+      final TekuConfiguration tekuConfig,
       final ServiceConfig serviceConfig,
-      ExecutionClientProvider executionClientProvider) {
+      final Optional<ExecutionClientProvider> executionClientProvider) {
     if (tekuConfig.beaconChain().interopConfig().isInteropEnabled()
-        || (!tekuConfig.powchain().isEnabled() && executionClientProvider.getWeb3j().isEmpty())) {
+        || (!tekuConfig.powchain().isEnabled() && executionClientProvider.isEmpty())) {
       return Optional.empty();
     }
 
     return Optional.of(
-        new PowchainService(
-            serviceConfig,
-            tekuConfig.powchain(),
-            executionClientProvider.getWeb3j(),
-            executionClientProvider.getEndpoint()));
+        new PowchainService(serviceConfig, tekuConfig.powchain(), executionClientProvider));
   }
 }