+++ CHANGELOG.md
@@ -19,10 +19,10 @@ For information on changes in released versions of Teku, see the [releases page]
 
 ### Additions and Improvements
 - Improved performance when regenerating non-finalized states that had to be dropped from memory.
-- Performance optimizations for Gnosis beacon chain
 - Performance optimizations for Gnosis beacon chain.
 - Improved performance when processing epoch transitions.
 - Added `is_optimistic` field to `/eth/v1/node/syncing` response.
+- Using execution engine endpoint as Eth1 endpoint when latter is not provided 
 
 ### Bug Fixes
 - Added stricter limits on attestation pool size.